diversifix_server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
astroid==3.0.0
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dill==0.3.7
emoji==2.8.0
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
fsspec==2023.9.2
gunicorn==21.2.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy==1.5.1
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.0
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
protobuf==4.24.4
pylint==3.0.1
python-dotenv==1.0.0
requests==2.31.0
split-words==0.1.3
stanza==1.6.1
sympy==1.12
tomli==2.0.1
tomlkit==0.12.1
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
yarl==1.9.2
